Upload the ObjectPooler script for the first time.
Unity description: You may have made games in the past that created and destroyed a lot of items (such as bullets or enemies). What you may not have known is that the act of instantiating and destroying are inefficient and can slow your projects down. This way to optimize our project is called object pooling.